DteFixers Gridpane converted to TilePane
