+++ src/main/java/com/girbola/controllers/datefixer/CheckBoxCell_Cameras.java
@@ -87,7 +87,7 @@ public class CheckBoxCell_Cameras extends TableCell<EXIF_Data_Selector, Boolean>
                             }
                             int counter = 0;
                             if (newValue == true) {
-                                for (Node node : model_DateFix.getGridPane().getChildren()) {
+                                for (Node node : model_DateFix.getTilePane().getChildren()) {
                                     if (node instanceof VBox && node.getId().equals("imageFrame")) {
                                         FileInfo fi = (FileInfo) node.getUserData();
                                         if (fi.getCamera_model() == null) {
@@ -104,7 +104,7 @@ public class CheckBoxCell_Cameras extends TableCell<EXIF_Data_Selector, Boolean>
                                     }
                                 }
                             } else if (newValue == false) {
-                                for (Node node : model_DateFix.getGridPane().getChildren()) {
+                                for (Node node : model_DateFix.getTilePane().getChildren()) {
                                     if (node instanceof VBox && node.getId().equals("imageFrame")) {
                                         FileInfo fi = (FileInfo) node.getUserData();
                                         if (fi.getCamera_model() == null) {
