+++ src/main/java/com/girbola/controllers/datefixer/CheckBoxCell_Events.java
@@ -95,7 +95,7 @@ public class CheckBoxCell_Events extends TableCell<EXIF_Data_Selector, Boolean> 
 							}
 							int counter = 0;
 							if (newValue == true) {
-								for (Node node : model_DateFix.getGridPane().getChildren()) {
+								for (Node node : model_DateFix.getTilePane().getChildren()) {
 									if (node instanceof VBox && node.getId().equals("imageFrame")) {
 
 										Messages.sprintf("NODEEEE: " + node.getId());
@@ -111,7 +111,7 @@ public class CheckBoxCell_Events extends TableCell<EXIF_Data_Selector, Boolean> 
 									}
 								}
 							} else if (newValue == false) {
-								for (Node node : model_DateFix.getGridPane().getChildren()) {
+								for (Node node : model_DateFix.getTilePane().getChildren()) {
 									if (node instanceof VBox && node.getId().equals("imageFrame")) {
 										FileInfo fi = (FileInfo) node.getUserData();
 										if (has_events(fi.getEvent(), listOfEvents)) {
