+++ src/main/java/com/girbola/controllers/datefixer/Model_datefix.java
@@ -47,7 +47,6 @@ import java.awt.image.BufferedImage;
 import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.sql.Connection;
@@ -110,7 +109,7 @@ public class Model_datefix extends DateFixerModel {
 		this.tilePane = tilePane;
 	}
 
-	private GridPane gridPane;
+	//private GridPane gridPane;
 	private TableView<EXIF_Data_Selector> cameras_TableView;
 	private TableView<EXIF_Data_Selector> dates_TableView;
 	private TableView<EXIF_Data_Selector> events_TableView;
@@ -262,13 +261,13 @@ public class Model_datefix extends DateFixerModel {
 		this.quickPick_tilePane = quickPick_tilePane;
 	}
 
-	public GridPane getGridPane() {
-		return gridPane;
-	}
-
-	public void setGridPane(GridPane gridPane) {
-		this.gridPane = gridPane;
-	}
+//	public GridPane getGridPane() {
+//		return gridPane;
+//	}
+//
+//	public void setGridPane(GridPane gridPane) {
+//		this.gridPane = gridPane;
+//	}
 
 	public FolderInfo getFolderInfo_full() {
 		return folderInfo_full;
@@ -429,22 +428,22 @@ public class Model_datefix extends DateFixerModel {
 		updateLocationInfos(fileInfo_List);
 	}
 
-	public void updateAllInfos(GridPane gridPane) {
+	public void updateAllInfos(TilePane gridPane) {
 		getCameras_TableView().getItems().clear();
 		getDates_TableView().getItems().clear();
 		getEvents_TableView().getItems().clear();
 		getLocations_TableView().getItems().clear();
 
-		updateCameraInfo(gridPane);
-		updateDatesInfos(gridPane);
-		updateEventsInfos(gridPane);
-		updateLocationsInfos(gridPane);
+		updateCameraInfo(tilePane);
+		updateDatesInfos(tilePane);
+		updateEventsInfos(tilePane);
+		updateLocationsInfos(tilePane);
 	}
 
-	public void updateCameraInfo(GridPane gridPane) {
+	public void updateCameraInfo(TilePane tilePane) {
 
 		List<FileInfo> fileInfo_list = new ArrayList<>();
-		for (Node node : gridPane.getChildren()) {
+		for (Node node : tilePane.getChildren()) {
 			if (node instanceof VBox) {
 				FileInfo fileInfo = (FileInfo) node.getUserData();
 				if (fileInfo != null) {
@@ -460,9 +459,9 @@ public class Model_datefix extends DateFixerModel {
 		}
 	}
 
-	public void updateEventsInfos(GridPane gridPane) {
+	public void updateEventsInfos(TilePane tilePane) {
 		List<FileInfo> fileInfo_list = new ArrayList<>();
-		for (Node node : gridPane.getChildren()) {
+		for (Node node : tilePane.getChildren()) {
 			if (node instanceof VBox) {
 				FileInfo fileInfo = (FileInfo) node.getUserData();
 				if (fileInfo != null) {
@@ -474,9 +473,9 @@ public class Model_datefix extends DateFixerModel {
 				Field.EVENT.getType());
 	}
 
-	public void updateLocationsInfos(GridPane gridPane) {
+	public void updateLocationsInfos(TilePane tilePane) {
 		List<FileInfo> fileInfo_list = new ArrayList<>();
-		for (Node node : gridPane.getChildren()) {
+		for (Node node : tilePane.getChildren()) {
 			if (node instanceof VBox) {
 				FileInfo fileInfo = (FileInfo) node.getUserData();
 				if (fileInfo != null) {
@@ -488,9 +487,9 @@ public class Model_datefix extends DateFixerModel {
 				Field.LOCATION.getType());
 	}
 
-	public void updateDatesInfos(GridPane gridPane) {
+	public void updateDatesInfos(TilePane tilePane) {
 		List<FileInfo> fileInfo_list = new ArrayList<>();
-		for (Node node : gridPane.getChildren()) {
+		for (Node node : tilePane.getChildren()) {
 			if (node instanceof VBox && node.getId().equals("imageFrame")) {
 				FileInfo fileInfo = (FileInfo) node.getUserData();
 				if (fileInfo != null) {
@@ -504,7 +503,7 @@ public class Model_datefix extends DateFixerModel {
 		dateFix_Utils.createDates_list(fileInfo_list);
 	}
 
-	public void acceptEverything() {
+	public void acceptEverything(TilePane tilePane) {
 		boolean changed = false;
 		// CssStylesController css = new CssStylesController();
 		Dialog<ButtonType> changesDialog = Dialogs.createDialog_YesNo(
@@ -513,7 +512,7 @@ public class Model_datefix extends DateFixerModel {
 
 		Optional<ButtonType> result = changesDialog.showAndWait();
 		if (result.get().getButtonData().equals(ButtonBar.ButtonData.YES)) {
-			for (Node node : getGridPane().getChildren()) {
+			for (Node node : tilePane.getChildren()) {
 				TextField tf = getTextField(node);
 				if (tf != null) {
 					if (!tf.getStyle().equals(CssStylesController.getBad_style())) {
@@ -681,11 +680,11 @@ public class Model_datefix extends DateFixerModel {
 		}
 	}
 
-	public void exitDateFixerWindow(GridPane gridPane, Window owner, WindowEvent event) {
+	public void exitDateFixerWindow(TilePane tilePane, Window owner, WindowEvent event) {
 		Messages.sprintf("exitDateFixerWindow");
 		model_Main.getMonitorExternalDriveConnectivity().cancel();
 
-		int badDates = checkIfRedDates(gridPane);
+		int badDates = checkIfRedDates(tilePane);
 		if (badDates != 0) {
 			Dialog<ButtonType> dialog = Dialogs.createDialog_YesNoCancel(owner,
 					bundle.getString("badFilesFoundWantToClose"));
@@ -697,7 +696,7 @@ public class Model_datefix extends DateFixerModel {
 			if (result.get().getButtonData().equals(ButtonBar.ButtonData.YES)) {
 				Stage stage = (Stage) Main.scene_Switcher.getScene_dateFixer().getWindow();
 				stage.setScene(Main.scene_Switcher.getScene_dateFixer());
-				saveThumbs();
+				saveThumbs(tilePane);
 				event.consume();
 
 			} else if (result.get().getButtonData().equals(ButtonBar.ButtonData.CANCEL_CLOSE)) {
@@ -707,7 +706,7 @@ public class Model_datefix extends DateFixerModel {
 		}
 
 		if (Main.conf.isSavingThumb()) {
-			saveThumbs();
+			saveThumbs(tilePane);
 		}
 
 		if (changes_made.get()) {
@@ -723,7 +722,7 @@ public class Model_datefix extends DateFixerModel {
 			}
 			Optional<ButtonType> result = changesDialog.showAndWait();
 			if (result.get().getButtonData().equals(ButtonBar.ButtonData.YES)) {
-				acceptEverything();
+				acceptEverything(tilePane);
 				Stage stage = (Stage) Main.scene_Switcher.getScene_dateFixer().getWindow();
 				stage.setScene(Main.scene_Switcher.getScene_dateFixer());
 				Main.scene_Switcher.getWindow().setOnCloseRequest(model_Main.exitProgram);
@@ -768,7 +767,7 @@ public class Model_datefix extends DateFixerModel {
 
 	}
 
-	public void saveThumbs() {
+	public void saveThumbs(TilePane tilePane) {
 		if (!Main.conf.isSavingThumb()) {
 			Messages.sprintf("isSavingThumb() were turned off");
 			return;
@@ -777,7 +776,7 @@ public class Model_datefix extends DateFixerModel {
 		getRenderVisibleNode().stopTimeLine();
 		ConcurrencyUtils.stopExecThreadNow();
 		List<ThumbInfo> thumbInfo_list = new ArrayList<>();
-		for (Node n : getGridPane().getChildren()) {
+		for (Node n : tilePane.getChildren()) {
 			if (n instanceof VBox) {
 				for (Node vbox : ((VBox) n).getChildren()) {
 					if (vbox instanceof StackPane) {
@@ -1025,9 +1024,9 @@ public class Model_datefix extends DateFixerModel {
 		return connection;
 	}
 
-	private int checkIfRedDates(GridPane gridPane) {
+	private int checkIfRedDates(TilePane tilePane) {
 		int counter = 0;
-		for (Node n : gridPane.getChildren()) {
+		for (Node n : tilePane.getChildren()) {
 			if (n instanceof VBox) {
 				if (n.getId().contains("imageFrame")) {
 					for (Node vbox : ((VBox) n).getChildren()) {
