+++ src/main/java/com/girbola/controllers/datefixer/RenderVisibleNode.java
@@ -258,17 +258,11 @@ public class RenderVisibleNode {
 					exec_single.submit(slow_render);
 				}
 			}
-
 			exec_single.shutdown();
 		} else {
 			sprintf("Visible node were empty");
 		}
 	}
-	//
-	// private int handle_tiff_thumb(FileInfo fileInfo, boolean betterQualityThumbs)
-	// {
-	// return 0;
-	// }
 
 	/**
 	 * 0 = thumbinfo found 1 = thumbinfo create thumbs
@@ -307,9 +301,9 @@ public class RenderVisibleNode {
 
 		Node mainNode = scrollPane.getContent();
 
-		if (mainNode instanceof GridPane && mainNode.getId().equals("dateFixer")) {
+		if (mainNode instanceof TilePane && mainNode.getId().equals("dateFixer")) {
 			Messages.sprintf("datefixer");
-			for (Node gridPane : ((GridPane) mainNode).getChildren()) {
+			for (Node gridPane : ((TilePane) mainNode).getChildren()) {
 				if (gridPane instanceof VBox && gridPane.getId().equals("imageFrame")) {
 					Bounds nodeBounds = gridPane.localToScene(gridPane.getBoundsInLocal());
 					if (paneBounds.intersects(nodeBounds)) {
